+++ src/Components/Header/index.tsx
@@ -24,6 +24,11 @@ const HeaderComponent = ({toggleTheme}: PropsHeader)=>{
             </Link>
           </li>
           <li onClick={ () => setShowMenu(false) }>
+            <Link href={"/blog"}>
+              <a>Blog</a>
+            </Link>
+          </li>
+          <li onClick={ () => setShowMenu(false) }>
             <Link href={"/Sobre"}>
               <a>Sobre</a>
             </Link>
